// modules/text/domain/enums/options-text.ts
export enum PromotingText {
    PRODUCT = 'Product',
    SERVICE = 'Service',
    APP = 'App',
    COURSE = 'Course',
    PERSONAL_BRAND = 'Personal brand',
    OTHER = 'Other'
}

export enum ToneText {
    PROFESSIONAL = 'Professional',
    EMOTIONAL = 'Emotional',
    FUN = 'Fun',
    CREATIVE = 'Creative',
    INFORMAL = 'Informal'
}

export enum LengthText {
    SHORT = 'Short',
    MEDIUM = 'Medium',
    LARGE = 'Large'
}

export enum FormatText {
    DESCRIPTION = 'Description',
    NETWORK_ANNOUNCEMENT = 'Network Announcement',
    IMPACT_PHRASE = 'Impact phrase',
    SLOGAN = 'Slogan'
}